Added Lukas Novotny to Translator Credits (Czech)
2006-01-05  Aaron Bockover  <aaron@aaronbock.net>

    * src/AboutBox.cs: Added Lukas Novotny to Translator Credits (Czech)

    * src/Banshee.Base/Library.cs: Allow single, collection, and queued
    removal of tracks based on URI as well as LibraryTrackInfo object reference

    * src/Banshee.Plugins/FileSystemMonitor/Watcher.cs: Implement removal
    and importing of files using a threaded queue processor instead of
    looping in an always-running thread to 'listen' for changes on
    two ArrayLists; removal and importing updated to use new mechanisims in
    Banshee.Base for interacting with the Library

    * src/Banshee.Plugins/FileSystemMonitor/Watch.cs: Simplified many aspects
    of the base Watch class; use events for notification instead of collections
    on which a thread 'listens'

    * src/Banshee.Plugins/FileSystemMonitor/InotifyWatch.cs:
    * src/Banshee.Plugins/FileSystemMonitor/FileSystemWatcherWatch.cs:
    Updated to implement changes in updated and simplified base class

+++ src/Banshee.Base/Library.cs
@@ -191,21 +191,20 @@ namespace Banshee.Base
                 TracksFnKeyed.Remove(MakeFilenameKey(track.Uri));
             }
         }
-        
-        /*private void CollectionRemove(int trackID, System.Uri trackUri)
+
+        private void Remove(Uri trackUri)
         {
-            lock(Tracks.SyncRoot) {
-                Tracks.Remove(trackID);
-            }
-            
-            lock(TracksFnKeyed.SyncRoot) {
-                TracksFnKeyed.Remove(MakeFilenameKey(trackUri));
-            }
-        }*/
+            Remove(TracksFnKeyed[MakeFilenameKey(trackUri)] as LibraryTrackInfo);
+        }
         
         public void Remove(LibraryTrackInfo track)
         {
+            if(track == null) {
+                return;
+            }
+            
             CollectionRemove(track);
+            
             Db.Execute(String.Format(
                 @"DELETE FROM Tracks
                     WHERE TrackID = '{0}'",
@@ -228,20 +227,26 @@ namespace Banshee.Base
             int remove_count = 0;
             int invalid_count = 0;
             
-            foreach(TrackInfo track in tracks) {
-                if(!(track is LibraryTrackInfo)) {
+            foreach(object o in tracks) {
+                LibraryTrackInfo track = null;
+                
+                if(o is Uri) {
+                    track = TracksFnKeyed[MakeFilenameKey(o as Uri)] as LibraryTrackInfo;
+                } else if(o is LibraryTrackInfo) {
+                    track = o as LibraryTrackInfo;
+                } 
+                
+                if(track == null) {
                     invalid_count++;
                     continue;
                 }
-                
-                LibraryTrackInfo library_track = track as LibraryTrackInfo;
-                
-                query += String.Format(" TrackID = '{0}' ", library_track.TrackId);
+
+                query += String.Format(" TrackID = '{0}' ", track.TrackId);
                 if(remove_count < tracks.Count - invalid_count - 1) {
                     query += " OR ";
                 }
                 
-                CollectionRemove(library_track);
+                CollectionRemove(track);
                 remove_count++;
             }
             
@@ -264,6 +269,11 @@ namespace Banshee.Base
         public void QueueRemove(TrackInfo track)
         {
             remove_queue.Add(track);
+        }
+        
+        public void QueueRemove(Uri trackUri)
+        {
+            remove_queue.Add(trackUri);
         }
         
         public void CommitRemoveQueue()
